% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_similarity_matrix.R
\name{kmeRs_generate_kmers}
\alias{kmeRs_generate_kmers}
\alias{kmeRs_similarity_matrix}
\title{Pairwise Similarity Matrix}
\usage{
kmeRs_generate_kmers(k, bases)
}
\arguments{
\item{k}{length of k-mers to calculate the similarity matrix for, defaults to 3;
e.g. for DNA, N = 4^3 = 64 combinations if code{k = 3};}

\item{q}{query vector with given k-mers}

\item{x}{kmers to search the query vector against. If unspecified, code{q} will 
be compared to either other k-mers within code{q} (code{compare.all = FALSE}),
or all possible combinations specified by the parameter code{k}}

\item{seq.type}{type of sequence in question, either 'DNA' or 'AA' (default);
this will also modify code{q} accordingly, if code{q} is unspecified.}

\item{compare.all}{if code{TRUE}, the query vector will be compared to all
possible combinations of k-mers (defaults to code{FALSE})}

\item{align.type}{type of alignment, either code{global} or {local}.
code{global} uses Needleman-Wunsch global alignment to calculate scores, while
code{local} represents Smith-Waterman local alignment instead}

\item{submat}{substitution matrix, default to 'BLOSUM62'; other choices are
'BLOSUM45', 'BLOSUM50', 'BLOSUM62', 'BLOSUM80', 'BLOSUM100', 'PAM30',
'PAM40', 'PAM70', 'PAM120' or 'PAM250'}

\item{save_to_file}{if specified, the results will be saved to the path in
comma-separated format (.CSV)}

\item{...}{other parameters, e.g. gap opening/extension penalties (code{gapOpening},
code{gapExtension}), or DNA match/mismatch scores (code{na.match}, code{na.mismatch})}
}
\value{
similarity matrix is returned as a data.frame
}
\description{
The \code{kmeRs_similarity_matrix} function generates a pairwise similarity score
matrix for for k length given k-mers vs. all possible k-mers combination.
The pairwise similarity score is calculated using PAM or BLOSUM substitution matrix;
30, 40, 70, 120, 250 and 62, 45, 50, 62, 80, 100 matrix versions are available for
PAM or BLOSUM, respectively. The results are evaluated by global similarity score;
higher similarity score indicates more similar sequences for BLOSUM and opposite for
PAM matrix.
}
\examples{
# this should return the standard BLOSUM62 matrix
kmeRs_similarity_matrix() 

# this should return a 4x4 identity matrix for DNA bases
kmeRs_similarity_matrix(seq.type = "DNA", na.match = 1, na.mismatch = 0)

}
