% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_heatmap.R
\name{kmeRs_heatmap}
\alias{kmeRs_heatmap}
\title{Generate a k-mer similarity score heatmap}
\usage{
kmeRs_heatmap(x, cexRow = NULL, cexCol = NULL, col = NULL)
}
\arguments{
\item{kmeRs_similarity_matrix}{matrix calculated by \code{kmeRs_similarity_matrix} function}

\item{summary_statistics_only}{when parameter is set to TRUE only the summarized table with
statistics is returned}
}
\value{
data.frame with results
}
\description{
The \code{kmeRs_heatmap} function makes a heatmap (placeholder ATM)
}
\examples{
# Use RColorBrewer to generate a figure similar to publication
library(RColorBrewer)
h.palette <- rev(brewer.pal(9, "YlGnBu"))
q0 <- c("GATTACA", "ACAGATT", "GAATTAC", "GAAATCT", "CTATAGA", "GTACATA", "AACGATT")
example <- kmeRs_similarity_matrix(q0, submat = "BLOSUM62")
kmeRs_heatmap(kmeRs_score(example), col = h.palette)

}
