% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_test_package.R
\name{kmeRs_test_package}
\alias{kmeRs_test_package}
\title{Simple Demo For KmeRs Package}
\usage{
kmeRs_test_package()
}
\value{
example report
}
\description{
The \code{kmeRs_test_package} function calculates and shows an example report of
the kmeRs package for sample given k-mers: "ATA", "CGC", "TGC", "GGA"
}
\examples{
# Test package - example raport

kmeRs_test_package()

}
